+++ ddsc/core/localstore.py
@@ -1,9 +1,9 @@
-import os
-import math
-import datetime
 import hashlib
+import math
 import mimetypes
-from ddsc.util import KindType
+import os
+
+from ddsc.core.util import KindType
 
 
 class LocalProject(object):
@@ -309,290 +309,12 @@ class HashUtil(object):
         return "md5", self.hash.hexdigest()
 
 
-class LocalOnlyCounter(object):
-    """
-    Visitor that counts items that need to be sent in LocalContent.
-    """
-    def __init__(self):
-        self.projects = 0
-        self.folders = 0
-        self.files = 0
-
-    def walk_project(self, project):
-        """
-        Increment counters for each project, folder, and files calling visit methods below.
-        :param project: LocalProject project we will count items of.
-        """
-        # This method will call visit_project, visit_folder, and visit_file below as it walks the project tree.
-        ProjectWalker.walk_project(project, self)
-
-    def visit_project(self, item):
-        """
-        Increments counter if the project is not already remote.
-        :param item: LocalProject
-        """
-        if not item.remote_id:
-            self.projects += 1
-
-    def visit_folder(self, item, parent):
-        """
-        Increments counter if item is not already remote
-        :param item: LocalFolder
-        :param parent: LocalFolder/LocalProject
-        """
-        if not item.remote_id:
-            self.folders += 1
-
-    def visit_file(self, item, parent):
-        """
-        Increments counter if item needs to be sent.
-        :param item: LocalFile
-        :param parent: LocalFolder/LocalProject
-        """
-        if item.need_to_send:
-            self.files += 1
-
-    def total_items(self):
-        """
-        Total number of items that need to be sent.
-        :return: int number of items to be sent.
-        """
-        return self.projects + self.folders + self.files
-
-    def result_str(self):
-        """
-        Return a string representing the totals contained herein.
-        :return: str counts/types string
-        """
-        return '{}, {}, {}'.format(LocalOnlyCounter.plural_fmt('project', self.projects),
-                                   LocalOnlyCounter.plural_fmt('folder', self.folders),
-                                   LocalOnlyCounter.plural_fmt('file', self.files))
-
-    @staticmethod
-    def plural_fmt(name, cnt):
-        """
-        pluralize name if necessary and combine with cnt
-        :param name: str name of the item type
-        :param cnt: int number items of this type
-        :return: str name and cnt joined
-        """
-        if cnt == 1:
-            return '{} {}'.format(cnt, name)
-        else:
-            return '{} {}s'.format(cnt, name)
-
-
-class UploadReport(object):
-    """
-    Creates a text report of items that were sent to the remote store.
-    """
-    def __init__(self, project_name):
-        """
-        Create report witht the specified project name since the local store doesn't contain that info.
-        :param project_name: str project name for the report
-        """
-        self.report_items = []
-        self.project_name = project_name
-        self._add_report_item('SENT FILENAME', 'ID', 'SIZE', 'HASH')
-
-    def _add_report_item(self, name, remote_id, size='', file_hash=''):
-        self.report_items.append(ReportItem(name, remote_id, size, file_hash))
-
-    def walk_project(self, project):
-        """
-        Create report items for each project, folder, and files if necessary calling visit_* methods below.
-        :param project: LocalProject project we will count items of.
-        """
-        # This method will call visit_project, visit_folder, and visit_file below as it walks the project tree.
-        ProjectWalker.walk_project(project, self)
-
-    def visit_project(self, item):
-        """
-        Add project to the report if it was sent.
-        :param item: LocalContent project level item
-        """
-        if item.sent_to_remote:
-            self._add_report_item('Project', item.remote_id)
-
-    def visit_folder(self, item, parent):
-        """
-        Add folder to the report if it was sent.
-        :param item: LocalFolder folder to possibly add
-        :param parent: LocalFolder/LocalContent not used here
-        """
-        if item.sent_to_remote:
-            self._add_report_item(item.path, item.remote_id)
-
-    def visit_file(self, item, parent):
-        """
-        Add file to the report if it was sent.
-        :param item: LocalFile file to possibly add.
-        :param parent: LocalFolder/LocalContent not used here
-        """
-        if item.sent_to_remote:
-            (alg, file_hash) = item.get_hashpair()
-            self._add_report_item(item.path, item.remote_id, item.size, file_hash)
-
-    def _report_header(self):
-        return u"Upload Report for Project: '{}' {}\n".format(self.project_name, datetime.datetime.utcnow())
-
-    def _report_body(self):
-        max_name = UploadReport.max_len([item.name for item in self.report_items])
-        max_remote_id = UploadReport.max_len([item.remote_id for item in self.report_items])
-        max_size = UploadReport.max_len([item.size for item in self.report_items])
-        return [item.str_with_sizes(max_name, max_remote_id, max_size) for item in self.report_items]
-
-    @staticmethod
-    def max_len(values):
-        return max([len(x) for x in values])
-
-    def get_content(self):
-        lines = [self._report_header()]
-        lines.extend(self._report_body())
-        return '\n'.join(lines)
-
-
-class ReportItem(object):
-    """
-    Item sent to remote store that is part of the UploadReport.
-    """
-    def __init__(self, name, remote_id, size='', file_hash=''):
-        """
-        Setup properties for use in str method
-        :param name: str name of the
-        :param remote_id: str remote uuid of the item
-        :param size: int/str size of the item can be '' if blank
-        :return:
-        """
-        self.name = name
-        self.remote_id = remote_id
-        self.size = str(size)
-        self.file_hash = file_hash
-
-    def str_with_sizes(self, max_name, max_remote_id, max_size):
-        """
-        Create string for report based on internal properties using sizes to line up columns.
-        :param max_name: int width of the name column
-        :param max_remote_id: int width of the remote_id column
-        :return: str info from this report item
-        """
-        name_str = self.name.ljust(max_name)
-        remote_id_str = self.remote_id.ljust(max_remote_id)
-        size_str = self.size.ljust(max_size)
-        return u'{}    {}    {}    {}'.format(name_str, remote_id_str, size_str, self.file_hash)
-
-
-class ProjectWalker(object):
-    @staticmethod
-    def walk_project(project, visitor):
-        """
-
-        :param project: LocalProject project we want to visit all children of.
-        :param visitor: object must implement visit_project, visit_folder, visit_file
-        :return:
-        """
-        ProjectWalker._visit_content(project, None, visitor)
-
-    @staticmethod
-    def _visit_content(item, parent, visitor):
-        """
-        Recursively visit nodes in the project tree.
-        :param item: LocalContent/LocalFolder/LocalFile we are traversing down from
-        :param parent: LocalContent/LocalFolder parent or None
-        :param visitor: object visiting the tree
-        """
-        if KindType.is_project(item):
-            visitor.visit_project(item)
-        elif KindType.is_folder(item):
-            visitor.visit_folder(item, parent)
-        else:
-            visitor.visit_file(item, parent)
-        if not KindType.is_file(item):
-            for child in item.children:
-                ProjectWalker._visit_content(child, item, visitor)
-
-
-class LocalOnlyCounter(object):
-    """
-    Visitor that counts items that need to be sent in LocalContent.
-    """
-    def __init__(self, bytes_per_chunk):
-        self.projects = 0
-        self.folders = 0
-        self.files = 0
-        self.chunks = 0
-        self.bytes_per_chunk = bytes_per_chunk
-
-    def walk_project(self, project):
-        """
-        Increment counters for each project, folder, and files calling visit methods below.
-        :param project: LocalProject project we will count items of.
-        """
-        # This method will call visit_project, visit_folder, and visit_file below as it walks the project tree.
-        ProjectWalker.walk_project(project, self)
-
-    def visit_project(self, item):
-        """
-        Increments counter if the project is not already remote.
-        :param item: LocalProject
-        """
-        if not item.remote_id:
-            self.projects += 1
-
-    def visit_folder(self, item, parent):
-        """
-        Increments counter if item is not already remote
-        :param item: LocalFolder
-        :param parent: LocalFolder/LocalProject
-        """
-        if not item.remote_id:
-            self.folders += 1
-
-    def visit_file(self, item, parent):
-        """
-        Increments counter if item needs to be sent.
-        :param item: LocalFile
-        :param parent: LocalFolder/LocalProject
-        """
-        if item.need_to_send:
-            self.files += 1
-            self.chunks += item.count_chunks(self.bytes_per_chunk)
-
-    def total_items(self):
-        """
-        Total number of files/folders/chunks that need to be sent.
-        :return: int number of items to be sent.
-        """
-        return self.projects + self.folders + self.chunks
-
-    def result_str(self):
-        """
-        Return a string representing the totals contained herein.
-        :return: str counts/types string
-        """
-        return '{}, {}, {}'.format(LocalOnlyCounter.plural_fmt('project', self.projects),
-                                   LocalOnlyCounter.plural_fmt('folder', self.folders),
-                                   LocalOnlyCounter.plural_fmt('file', self.files))
-
-    @staticmethod
-    def plural_fmt(name, cnt):
-        """
-        pluralize name if necessary and combine with cnt
-        :param name: str name of the item type
-        :param cnt: int number items of this type
-        :return: str name and cnt joined
-        """
-        if cnt == 1:
-            return '{} {}'.format(cnt, name)
-        else:
-            return '{} {}s'.format(cnt, name)
-
-
 class FileChunkReader(object):
     """
     Reads chunks of data from a file based on a index and a size and passes them to a processor.
     So we seek to index*chunk_size, then read one chunk at a time and give it to processor
     until we have done num_chunks_to_process.
+    NOT CURRENTLY USED! To be used for parallel upload.
     """
     def __init__(self, filename, chunk_size, index, num_chunks_to_process):
         """
