+++ ddsc/core/util.py
