+++ ddsc/ddsclient.py
@@ -1,12 +1,14 @@
+""" Runs the appropriate command for a user based on arguments. """
 from __future__ import print_function
+
 import datetime
 
-from ddsc.localstore import LocalProject, LocalOnlyCounter, UploadReport
-from ddsc.remotestore import RemoteStore, RemoteContentDownloader
-from ddsc.cmdparser import CommandParser, path_does_not_exist_or_is_empty, replace_invalid_path_chars
-from ddsc.util import ProgressPrinter
-from ddsc.handover import Handover
+from ddsc.core.handover import ProjectHandover
+from ddsc.core.remotestore import RemoteStore
+from ddsc.core.upload import ProjectUpload
 
+from ddsc.cmdparser import CommandParser, path_does_not_exist_or_is_empty, replace_invalid_path_chars
+from ddsc.core.download import ProjectDownload
 
 
 class DDSClient(object):
@@ -58,29 +60,22 @@ class DDSClient(object):
         :param command_constructor: class of an object that implements run(args)
         :param args: object arguments for specific command created by CommandParser
         """
-        command = command_constructor(self)
+        command = command_constructor(self.config)
         command.run(args)
-
-    def create_remote_store(self):
-        """
-        Create a remote store based on config.
-        :return: RemoteStore remote store based on config settings.
-        """
-        return RemoteStore(self.config)
 
 
 class UploadCommand(object):
     """
     Uploads a folder to a remote project.
     """
-    def __init__(self, parent):
+    def __init__(self, config):
         """
-        Pass in the parent who can create a remote_store/url so we can access the remote data.
-        :param parent: DDSClient parent who can create objects based on config for us.
+        Pass in the config containing remote_store/url so we can access the remote data.
+        :param config: Config global configuration for use with this command.
         """
-        self.remote_store = parent.create_remote_store()
-        self.base_url = parent.config.get_url_base()
-        self.config = parent.config
+        self.remote_store = RemoteStore(config)
+        self.base_url = config.get_url_base()
+        self.config = config
 
     def run(self, args):
         """
@@ -93,80 +88,26 @@ class UploadCommand(object):
         folders = args.folders                  # list of local files/folders to upload into the project
         follow_symlinks = args.follow_symlinks  # should we follow symlinks when traversing folders
 
-        remote_project = self.remote_store.fetch_remote_project(project_name)
-        local_project = self._load_local_project(folders, follow_symlinks)
-        local_project.update_remote_ids(remote_project)
-        different_items = self._count_differences(local_project)
-        self._print_differences_summary(different_items)
-        if different_items.total_items() != 0:
-            self._upload_differences(local_project, project_name, different_items.total_items())
-            self._print_report(project_name, local_project)
-        self._print_url(local_project)
-
-    def _load_local_project(self, folders, follow_symlinks):
-        local_project = LocalProject(followsymlinks=follow_symlinks)
-        local_project.add_paths(folders)
-        return local_project
-
-    def _count_differences(self, local_project):
-        """
-        Count how many things we will be sending.
-        :param local_project: LocalProject project we will send data from
-        :return: LocalOnlyCounter contains counts for various items
-        """
-        different_items = LocalOnlyCounter(self.config.upload_bytes_per_chunk)
-        different_items.walk_project(local_project)
-        return different_items
-
-    def _print_differences_summary(self, different_items):
-        """
-        Print a summary of what is to be done.
-        :param different_items: LocalOnlyCounter item that contains the summary
-        """
-        print('Uploading {}.'.format(different_items.result_str()))
-
-    def _upload_differences(self, local_project, project_name, different_items_cnt):
-        """
-        Send different items within local_project to remote store
-        :param local_project: LocalProject project we will send data from
-        :param different_items_cnt: int count of items to be sent for progress bar
-        """
-        progress_printer = ProgressPrinter(different_items_cnt, msg_verb='sending')
-        self.remote_store.upload_differences(local_project,
-                                             project_name,
-                                             progress_printer)
-        progress_printer.finished()
-
-    def _print_report(self, project_name, local_project):
-        """
-        Generate and print a report onto stdout.
-        """
-        report = UploadReport(project_name)
-        report.walk_project(local_project)
-        print('\n')
-        print(report.get_content())
-        print('\n')
-
-    def _print_url(self, local_project):
-        """
-        Print url to view the project via dds portal.
-        """
-        msg = 'URL to view project'
-        project_id = local_project.remote_id
-        url = '{}: https://{}/portal/#/project/{}'.format(msg, self.base_url, project_id)
-        print(url)
+        project_upload = ProjectUpload(self.config, project_name, folders, follow_symlinks=follow_symlinks)
+        print(project_upload.get_differences_summary())
+        if project_upload.needs_to_upload():
+            project_upload.run()
+            print('\n')
+            print(project_upload.get_upload_report())
+            print('\n')
+        print(project_upload.get_url_msg())
 
 
 class DownloadCommand(object):
     """
     Downloads the content from a remote project into a folder.
     """
-    def __init__(self, parent):
+    def __init__(self, config):
         """
-        Pass in the parent who can create a remote_store so we can access the remote data.
-        :param parent: DDSClient parent who can create objects based on config for us.
+        Pass in the config who can create a remote_store so we can access the remote data.
+        :param config: Config global configuration for use with this command.
         """
-        self.remote_store = parent.create_remote_store()
+        self.remote_store = RemoteStore(config)
 
     def run(self, args):
         """
@@ -179,21 +120,20 @@ class DownloadCommand(object):
         if not folder:
             fixed_path = replace_invalid_path_chars(project_name.replace(' ', '_'))
             folder = path_does_not_exist_or_is_empty(fixed_path)
-        remote_project = self.remote_store.fetch_remote_project(project_name, must_exist=True)
-        downloader = RemoteContentDownloader(self.remote_store, folder)
-        downloader.walk_project(remote_project)
+        project_download = ProjectDownload(self.remote_store, project_name, folder)
+        project_download.run()
 
 
 class AddUserCommand(object):
     """
     Adds a user to a pre-existing remote project.
     """
-    def __init__(self, parent):
+    def __init__(self, config):
         """
-        Pass in the parent who can create a remote_store so we can access the remote data.
-        :param parent: DDSClient parent who can create objects based on config for us.
+        Pass in the config who can create a remote_store so we can access the remote data.
+        :param config: Config global configuration for use with this command.
         """
-        self.remote_store = parent.create_remote_store()
+        self.remote_store = RemoteStore(config)
 
     def run(self, args):
         """
@@ -214,13 +154,13 @@ class MailDraftCommand(object):
     """
     Send email that draft project is ready for a user.
     """
-    def __init__(self, parent):
+    def __init__(self, config):
         """
-        Pass in the parent who can create a remote_store so we can access the remote data.
-        :param parent: DDSClient parent who can create objects based on config for us.
+        Pass in the config who can create a remote_store so we can access the remote data.
+        :param config: Config global configuration for use with this command.
         """
-        self.remote_store = parent.create_remote_store()
-        self.handover = Handover(parent.config, self.remote_store)
+        self.remote_store = RemoteStore(config)
+        self.project_handover = ProjectHandover(config, self.remote_store, print_func=print)
 
     def run(self, args):
         """
@@ -231,7 +171,7 @@ class MailDraftCommand(object):
         email = args.email                  # email of person to send email to
         username = args.username            # username of person to send email to, will be None if email is specified
         to_user = self.remote_store.lookup_user_by_email_or_username(email, username)
-        dest_email = self.handover.mail_draft(project_name, to_user)
+        dest_email = self.project_handover.mail_draft(project_name, to_user)
         print("Email draft sent to " + dest_email)
 
 
@@ -239,13 +179,13 @@ class HandoverCommand(object):
     """
     Send handover email that project is ready for a user to receive.
     """
-    def __init__(self, parent):
+    def __init__(self, config):
         """
-        Pass in the parent who can create a remote_store so we can access the remote data.
-        :param parent: DDSClient parent who can create objects based on config for us.
+        Pass in the config who can create a remote_store so we can access the remote data.
+        :param config: Config global configuration for use with this command.
         """
-        self.remote_store = parent.create_remote_store()
-        self.handover = Handover(parent.config, self.remote_store)
+        self.remote_store = RemoteStore(config)
+        self.project_handover = ProjectHandover(config, self.remote_store, print_func=print)
 
     def run(self, args):
         """
@@ -260,7 +200,7 @@ class HandoverCommand(object):
         if not skip_copy_project:
             new_project_name = self.get_new_project_name(project_name)
         to_user = self.remote_store.lookup_user_by_email_or_username(email, username)
-        dest_email = self.handover.handover(project_name, new_project_name, to_user)
+        dest_email = self.project_handover.handover(project_name, new_project_name, to_user)
         print("Handover message sent to " + dest_email)
 
     def get_new_project_name(self, project_name):
